Split legacy graph validation tests
